+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Dropdown and DropdownSuggest components
 
 Dropdown and dropdown-suggest have been updated. As they share common functionality, dropdown and dropdown-suggest now use a List decorator. This should not affect how you use either component.
+* Dropdown now filters results as you type. 
 
 ## Dialog Cancel button